requirements.txt
accelerate==1.7.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.8.1
attrs==25.3.0
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
cssselect==1.3.0
datasets==3.6.0
dill==0.3.8
Django==5.2.1
django-cors-headers==4.8.0
djangorestframework==3.16.1
docstring_parser==0.17.0
evaluate==0.4.3
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.3.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.183.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.117.0
google-cloud-bigquery==3.38.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.39.1
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.75.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.1.2
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface-hub==0.32.2
idna==3.10
jieba3k==0.35.1
Jinja2==3.1.6
joblib==1.5.1
langdetect==1.0.9
lxml==5.4.0
lxml_html_clean==0.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.4.4
multiprocess==0.70.16
networkx==3.5
newspaper3k==0.2.8
nltk==3.9.1
numpy==2.2.6
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
packaging==25.0
pandas==2.3.0
pillow==11.2.1
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
psycopg2-binary==2.9.10
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.9
pydantic_core==2.33.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
sentence-transformers==5.1.0
sentencepiece==0.2.1
seqeval==1.2.2
setuptools==80.9.0
sgmllib3k==1.0.0
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlparse==0.5.3
sympy==1.13.1
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.11.0
tinysegmenter==0.3
tldextract==5.3.0
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
transformers==4.52.4
triton==3.2.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
websockets==15.0.1
xxhash==3.5.0
yarl==1.20.1
